+++ app/admin/custom_products.rb
@@ -1,8 +1,9 @@
 ActiveAdmin.register CustomProduct do
   permit_params :name, :description, :user_id
 
-  remove_filter :image_attachment
-  remove_filter :image_blob
+  remove_filter :highlighted_image_id
+  remove_filter :images_attachments
+  remove_filter :images_blobs
   remove_filter :possession
   remove_filter :sub_categories
 end