+++ app/controllers/products_controller.rb
@@ -139,10 +139,11 @@ class ProductsController < ApplicationController
     end
 
     @public_possessions_with_image = @product.possessions
-                                             .includes([:image_attachment])
+                                             .includes([:images_attachments])
                                              .joins(:user)
                                              .where(user: { profile_visibility: user_signed_in? ? [1, 2] : 2 })
-                                             .select { |possession| possession.image.attached? }
+                                             .select { |possession| possession.images.attached? }
+                                             .map { |possession| PossessionPresenter.new(possession) }
 
     @contributors = ActiveRecord::Base.connection.execute("
       SELECT DISTINCT