+++ db/migrate/20250401134747_add_highlighted_image_to_custom_product.rb
@@ -1,0 +1,5 @@
+class AddHighlightedImageToCustomProduct < ActiveRecord::Migration[8.0]
+  def change
+    add_column :custom_products, :highlighted_image, :bigint
+  end
+end