+++ db/migrate/20250401164244_rename_highlighted_image_to_highlighted_image_id.rb
@@ -1,0 +1,5 @@
+class RenameHighlightedImageToHighlightedImageId < ActiveRecord::Migration[8.0]
+  def change
+    rename_column :custom_products, :highlighted_image, :highlighted_image_id
+  end
+end