+++ db/migrate/20250401170247_add_highlighted_image_id_to_possession.rb
@@ -1,0 +1,5 @@
+class AddHighlightedImageIdToPossession < ActiveRecord::Migration[8.0]
+  def change
+    add_column :possessions, :highlighted_image_id, :bigint
+  end
+end